sched/env: add tg_envc in task_group_s to avoid some loop
VELAPLATFO-5872

+++ sched/environ/env_findvar.c
@@ -83,9 +83,9 @@ static bool env_cmpname(const char *pszname, const char *peqname)
  *
  ****************************************************************************/
 
-int env_findvar(FAR struct task_group_s *group, FAR const char *pname)
+ssize_t env_findvar(FAR struct task_group_s *group, FAR const char *pname)
 {
-  int i;
+  ssize_t i;
 
   /* Verify input parameters */
